+++ assets/js/views/downloadFormView.js
@@ -55,12 +55,21 @@ export default class DownloadFormView {
             keyParameter : 'statecode',
             parseKey : getStateFromCounty
         });
-        return new PlaceInputView({
-            $container : $('#place'),
-            countryModel : countryModel,
-            stateModel : stateModel,
-            countyModel : countyModel
-        });
+        if(this.$form.attr('id') == 'paramsBasic'){
+            return new PlaceInputView({
+                $container : $('#placeBasic'),
+                countryModel : countryModel,
+                stateModel : stateModel,
+                countyModel : countyModel
+            });
+        }else{
+            return new PlaceInputView({
+                $container : $('#place'),
+                countryModel : countryModel,
+                stateModel : stateModel,
+                countyModel : countyModel
+            });
+        }
     }
 
     /*
@@ -83,10 +92,6 @@ export default class DownloadFormView {
             siteTypeModel : new CachedCodes({codes : 'sitetype'}),
             organizationModel : new CachedCodes({codes : 'organization'})
         });
-        const nldiView = new NldiView({
-            mapDivId : 'nldi-map',
-            input: 'nldi-url'
-        });
         const samplingParametersInputView = new SamplingParameterInputView({
             $container : this.$form.find('#sampling'),
             sampleMediaModel : new CachedCodes({codes: 'samplemedia'}),
@@ -97,7 +102,7 @@ export default class DownloadFormView {
             assemblageModel : new CachedCodes({codes: 'assemblage'})
         });
         this.dataDetailsView = new DataDetailsView({
-            $container : this.$form.find('#download-box-input-div'),
+            $container : this.$form.find('.download-box-input-div'),
             updateResultTypeAction : (resultType) => {
                 this.$form.attr('action', queryService.getFormUrl(resultType));
             }
@@ -129,9 +134,15 @@ export default class DownloadFormView {
         this.dataDetailsView.initialize();
         pointLocationInputView.initialize();
         boundingBoxInputView.initialize();
-        if (Config.NLDI_ENABLED) {
+
+        // Only create map for advanced form
+        if (Config.NLDI_ENABLED && this.$form.attr('id') == "params") {
+            const nldiView = new NldiView({
+                mapDivId : 'nldi-map',
+                input: 'nldi-url'
+            });
             nldiView.initialize(); 
-        } else {
+        } else if (this.$form.attr('id') == "params"){
             this.$form.find('#nldi-container').hide();
             this.$form.find('#nldi-map').hide();
         }
@@ -164,21 +175,45 @@ export default class DownloadFormView {
         });
 
         let $dataProviders = this.$form.find('#providers-select');
-        // Add click handler for reset button
-        this.$form.find('.reset-button').click(() => {
+        // Add click handler for clear parameters button
+        this.$form.find('.reset-params').click(() => {
+            $('#withinBasic').val('').trigger('change');
+            $('#latBasic').val('').trigger('change');
+            $('#longBasic').val('').trigger('change');
+            $('#countrycodeBasic').val(null).trigger('change');
+            $('#statecodeBasic').val(null).trigger('change');
+            $('#countycodeBasic').val(null).trigger('change');
+            $('#siteTypeBasic').val(null).trigger('change');
+            placeInputView.resetContainer();
+            pointLocationInputView.resetContainer();
+            siteParameterInputView.resetContainer();
+        });
+
+        // Add click handler for clear filters button
+        this.$form.find('.reset-filters').click(() => {
+            $('#startDateLoBasic').val('').trigger('change');
+            $('#startDateHiBasic').val('').trigger('change');
+            let $checkboxes = $('.datasources-basic').find(':input');
+            $checkboxes.each(function(checkbox){
+                $checkboxes[checkbox].checked = true;
+            });
+            $('#siteCodeBasic').val(null).trigger('change');
+            $('#sampleMediaBasic').val(null).trigger('change');
+            $('#charGroupBasic').val(null).trigger('change');
+            samplingParametersInputView.resetContainer();
+        });
+
+        this.$form.find('#startOver').click(() => {
             placeInputView.resetContainer();
             pointLocationInputView.resetContainer();
             boundingBoxInputView.resetContainer();
             samplingParametersInputView.resetContainer();
-            biologicalSamplingInputView.resetContainer();
             siteParameterInputView.resetContainer();
             this.dataDetailsView.resetContainer();
-            $dataProviders.val('');
-            $dataProviders.trigger('change');
         });
 
         // Set up the Download button
-        this.$form.find('#main-button').click((event) => {
+        this.$form.find('.main-button').click((event) => {
             const fileFormat = this.dataDetailsView.getMimeType();
             const resultType = this.dataDetailsView.getResultType();
             const queryParamArray = this.getQueryParamArray();
@@ -198,6 +233,7 @@ export default class DownloadFormView {
             event.preventDefault();
 
             if (!downloadFormController.validateDownloadForm(this.$form)) {
+                console.log("invalid")
                 return;
             }